+++ .gitignore
@@ -15,9 +15,6 @@ web-build/
 assets/portfolio-assets.sketch
 
 /dist
-# @generated expo-cli sync-e8131d6da33d05e44f5bdcb96ba7e70722c4e5e4
-# The following patterns were generated by expo-cli
 
-expo-env.d.ts
-expo-env.d.ts
-# @end expo-cli
+
+src/expo-oss-showcase-data.json
