+++ src/app/games.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import ProjectCard from '../components/card/Card';
-import Layout from '../components/Layout';
-import PageHeader from '../components/PageHeader';
-import { Projects } from '../Data';
+import ProjectCard from '@/components/card/Card';
+import Layout from '@/components/Layout';
+import PageHeader from '@/components/PageHeader';
+import { Projects } from '@/Data';
 
 export default function Games() {
   return (
