+++ src/app/index.tsx
@@ -1,0 +1,1 @@
+export { default } from '@/app/talks';
