+++ src/components/AspectImage.tsx
