+++ src/constants/Quotes.ts
